fix: move component export assembly to proper folders
