+++ src/Tabs/deprecated/Tabs.scss
