+++ src/index.scss
@@ -8,4 +8,4 @@
 @import './Table/Table.scss';
 @import './TransitionReplace/TransitionReplace.scss';
 @import './ValidationMessage/ValidationMessage.scss';
-@import "./Tabs/Tabs.scss";
+@import "./Tabs/deprecated/Tabs.scss";
